Create README.md
This project was accomplished by firstly collecting vital data from the dual patient (1 and 2). The patient (1) is connected to bio-medical sensor such as body temperature, heart rate, IV Fluid and Piezo-electric into a microcontroller-based board such as Arduino Uno. The patient (2) is connected to bio-medical sensor such as body temperature, heart rate, Blood pressure and gas sensor into a microcontroller-based board such as Arduino Uno. Here the processing and calculation is done after which, secondly, the data is pulled to raspberry pi and it is transferred via a wireless-technology module to a physician. The wireless technology used for this project was using IoT, where the data is automatically and periodically are been stored and analyzed through web server or mobile App graphically. For patient (1) when the vital value goes down the threshold value a instant alert message is automatically generated from GSM to doctor. For patient (2) when the doctor arrives close range to a patient, the vital values are displayed over ArduTooth mobile App. A LCD is available in the both patient health care system so that instant view of the vital data can be seen.